Assignment-1/requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-google-genai==2.0.8
google-generativeai==0.8.4
faiss-cpu==1.9.0.post1
pypdf==5.1.0
sentence-transformers==3.3.1
streamlit==1.41.1
python-dotenv==1.0.1
ipykernel==6.29.5
